Add baseline profile module
Created with the baseline profile module wizard in IntelliJ. I additionally went through and made a quick test that allowed the generator to invoke the main screens of the app. This excludes the "Verify APK File" screen because I can't guarantee an APK will be somewhere on the virtual managed device.